pyproject.toml
[project]
name = "realtime-chat"
version = "0.1.0"
description = "Slack-like real-time chat backend with FastAPI, Socket.IO, PostgreSQL, Redis, Celery"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-socketio>=5.12.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis[hiredis]>=5.2.0",
    "celery[redis]>=5.4.0",
    "PyJWT>=2.10.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.18",
    "Pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "python-socketio[asyncio_client]>=5.12.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers -v"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/migrations/*",
    "*/__init__.py",
    "src/adapters/persistence/database.py",
    "src/adapters/persistence/models/*",
    "src/adapters/persistence/repositories/*",
    "src/adapters/celery_tasks/file_processing_tasks.py",
]

[tool.coverage.report]
fail_under = 90
show_missing = true
